+++ lib/src/internal/layout/remove_item_by_id.dart
@@ -1,24 +1,16 @@
 import 'package:docking/src/layout/docking_layout.dart';
-import 'package:docking/src/layout/layout_modifier.dart';
+import 'package:docking/src/internal/layout/layout_modifier.dart';
+import 'package:meta/meta.dart';
 
-/// Removes [DockingItem] from this layout.
-class RemoveItem extends LayoutModifier {
-  RemoveItem({required this.itemToRemove});
+/// Removes [DockingItem] by id from this layout.
+@internal
+class RemoveItemById extends LayoutModifier {
+  RemoveItemById({required this.id});
 
-  final DockingItem itemToRemove;
-
-  @override
-  void validate(DockingLayout layout, DockingArea area) {
-    super.validate(layout, area);
-    if (area.layoutId != layout.id) {
-      throw ArgumentError(
-          'DockingArea belongs to another layout. Keep the layout in the state of your StatefulWidget.');
-    }
-  }
+  final dynamic id;
 
   @override
   DockingArea? newLayout(DockingLayout layout) {
-    validate(layout, itemToRemove);
     if (layout.root != null) {
       return _buildLayout(layout.root!);
     }
@@ -29,16 +21,16 @@ class RemoveItem extends LayoutModifier {
   DockingArea? _buildLayout(DockingArea area) {
     if (area is DockingItem) {
       DockingItem dockingItem = area;
-      if (dockingItem == itemToRemove) {
+      if (dockingItem.id == id) {
         return null;
       }
-      return DockingItem.clone(dockingItem);
+      return dockingItem;
     } else if (area is DockingTabs) {
       DockingTabs dockingTabs = area;
       List<DockingItem> children = [];
       dockingTabs.forEach((child) {
-        if (child != itemToRemove) {
-          children.add(DockingItem.clone(child));
+        if (child.id != id) {
+          children.add(child);
         }
       });
       if (children.length == 1) {
@@ -63,13 +55,9 @@ class RemoveItem extends LayoutModifier {
         return children.first;
       }
       if (area is DockingRow) {
-        DockingRow row = DockingRow(children);
-        row.weights = area.weights;
-        return row;
+        return DockingRow(children);
       } else if (area is DockingColumn) {
-        DockingColumn column = DockingColumn(children);
-        column.weights = area.weights;
-        return column;
+        return DockingColumn(children);
       }
       throw ArgumentError(
           'DockingArea class not recognized: ' + area.runtimeType.toString());