+++ test/utils.dart
@@ -2,14 +2,40 @@ import 'package:docking/docking.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-DockingItem dockingItem(String? name) {
-  return DockingItem(name: name, widget: Container());
+DockingItem dockingItem(String? name, {dynamic id}) {
+  return DockingItem(id: id, name: name, widget: Container());
+}
+
+void removeItemById(DockingLayout layout, List<dynamic> ids) {
+  List<DockingItem> itemsToDispose = [];
+  for (dynamic id in ids) {
+    DockingItem? item = layout.findDockingItem(id);
+    if (item != null) {
+      itemsToDispose.add(item);
+    }
+  }
+  List<DockingArea> areas = layout.layoutAreas();
+  layout.removeItemByIds(ids);
+  areas.forEach((area) {
+    bool areaIsDisposedItem = false;
+    for (DockingItem disposedItem in itemsToDispose) {
+      if (area == disposedItem) {
+        areaIsDisposedItem = true;
+        break;
+      }
+    }
+    if (area is DockingItem && !areaIsDisposedItem) {
+      testNonDisposedArea(area);
+    } else {
+      testDisposed(area);
+    }
+  });
 }
 
 void removeItem(DockingLayout layout, DockingItem item) {
   List<DockingArea> areas = layout.layoutAreas();
   layout.removeItem(item: item);
-  testDisposedList(areas);
+  testOldAreas(areas, disposedItem: item);
 }
 
 void moveItem(DockingLayout layout, DockingItem draggedItem,
@@ -19,14 +45,14 @@ void moveItem(DockingLayout layout, DockingItem draggedItem,
       draggedItem: draggedItem,
       targetArea: targetArea,
       dropPosition: dropPosition);
-  testDisposedList(areas);
+  testOldAreas(areas);
 }
 
 void addItemOnRoot(
     DockingLayout layout, DockingItem newItem, DropPosition dropPosition) {
   List<DockingArea> areas = layout.layoutAreas();
   layout.addItemOnRoot(newItem: newItem, dropPosition: dropPosition);
-  testDisposedList(areas);
+  testOldAreas(areas);
 }
 
 void addItemOn(DockingLayout layout, DockingItem newItem, DropArea targetArea,
@@ -34,7 +60,7 @@ void addItemOn(DockingLayout layout, DockingItem newItem, DropArea targetArea,
   List<DockingArea> areas = layout.layoutAreas();
   layout.addItemOn(
       newItem: newItem, targetArea: targetArea, dropPosition: dropPosition);
-  testDisposedList(areas);
+  testOldAreas(areas);
 }
 
 int _testAreasAttributes(DockingArea parent, bool value, int level, int index) {
@@ -66,8 +92,19 @@ void testHierarchy(DockingLayout layout, String hierarchy) {
   }
 }
 
-void testDisposedList(List<DockingArea> layoutAreas) {
-  layoutAreas.forEach((area) => testDisposed(area));
+void testNonDisposedArea(DockingArea area) {
+  expect(area.layoutId > -1, true, reason: 'has layoutId');
+  expect(area.index > -1, true, reason: 'has index');
+}
+
+void testOldAreas(List<DockingArea> layoutAreas, {DockingItem? disposedItem}) {
+  layoutAreas.forEach((area) {
+    if (area is DockingItem && area != disposedItem) {
+      testNonDisposedArea(area);
+    } else {
+      testDisposed(area);
+    }
+  });
 }
 
 void testDockingArea(DockingArea area,